Make imports more sane
Reorganize files so that imports make more sense. Make benchpress
package and benchpress.py that uses the benchpress package so that we
don't have to hackily run a script that is part of a package and the
resulting sys.path magic

Also fix imports in tests
