+++ benchpress/benchpress/lib/parser_factory.py
@@ -6,16 +6,12 @@
 # LICENSE file in the root directory of this source tree. An additional grant
 # of patent rights can be found in the PATENTS file in the same directory.
 
-from lib.hook import Hook
+from .factory import BaseFactory
+from .parser import Parser
 
+from benchpress.plugins.parsers import register_parsers
 
-class NoopHook(Hook):
-    """NoopHook is the default hook used if no other hooks are specified. As the
-    name suggests, it doesn't do anything.
-    """
+ParserFactory = BaseFactory(Parser)
 
-    def before_job(self, opts):
-        pass
-
-    def after_job(self, opts):
-        pass
+# register third-party parsers with the factory
+register_parsers(ParserFactory)
