+++ benchpress/benchpress/plugins/parsers/schbench.py
@@ -6,7 +6,7 @@
 # LICENSE file in the root directory of this source tree. An additional grant
 # of patent rights can be found in the PATENTS file in the same directory.
 
-from lib.parser import Parser
+from benchpress.lib.parser import Parser
 
 
 class SchbenchParser(Parser):
